Move Look2Mouse() in LateUpdate() message
When an animator is added as component and the entity's sprite is
changed via code, the animator overrides this.

When Look2Mouse() was called in FixedUpdate(), the animator overrode
the changes of the sprite. Which made the player sprites to not be
displayed accordingly. Calling the Look2Mouse() in LateUpdate() is a
workaround of this.

Signed-off-by: ivaylolivanov <ivaylolivanov95@gmail.com>